Move sendJwt middleware to expressmiddleware file
