+++ routes/api/utils/ExpressMiddleware.js
@@ -1,3 +1,6 @@
+const jwt = require('jsonwebtoken');
+const keys = require('../../../config/keys');
+
 const restrictAccess = function(restrictedAccessLevels = []) {
     return (req, res, next) => {
         const validAccessLevels = ['Super', 'Admin', 'Volunteer'];
@@ -28,7 +31,42 @@ const expressMiddleware = {
     },
     restrictAccess: restrictAccess,
     restrictVolunteer: restrictAccess(['Volunteer']),
-    restrictAdminVolunteer: restrictAccess(['Volunteer', 'Admin'])
+    restrictAdminVolunteer: restrictAccess(['Volunteer', 'Admin']),
+
+    /**
+     * An express middleware that sends a new jwt.
+     * @param {request} req API request
+     * @param {response} res API response
+     */
+    sendJwt(req, res) {
+        // Create JWT Payload, basically what we want to send in the response
+        const { user } = req;
+        const payload = {
+            id: user.id,
+            firstName: user.firstName,
+            lastName: user.lastName,
+            seniorCenterId: user.seniorCenterId,
+            accessLevel: user.accessLevel
+        };
+
+        // Sign token
+        return jwt.sign(
+            payload,
+            keys.jwtKey,
+            {
+                expiresIn: 39600 // 11 hours in seconds
+            },
+            (err, token) => {
+                if (err) {
+                    return res.status(400).json(err);
+                }
+
+                return res.json({
+                    token: 'Bearer ' + token
+                });
+            }
+        );
+    }
 };
 
 module.exports = expressMiddleware;
